+++ MovieWebsite.Persons/Controllers/PersonController.cs
@@ -13,7 +13,7 @@ namespace MovieWebsite.Persons.Controllers
     public class PersonController : Controller
     {
         [HttpGet("FetchPerson")]
-        public async Task<IActionResult> FetchPerson(long id)
+        public async Task<IActionResult> FetchPerson(int id)
         {
             await using var db = new PersonContext();
             var person = await db.Persons
@@ -37,8 +37,8 @@ namespace MovieWebsite.Persons.Controllers
         /// <returns>Returns id of added Person if success</returns>
         [Authorize()]
         [HttpPost("AddPerson")]
-        public async Task<IActionResult> AddPerson(string name, string originalName, long posterId, string description,
-            [FromQuery] HashSet<long> characters, [FromQuery] HashSet<long> movies)
+        public async Task<IActionResult> AddPerson(string name, string originalName, int posterId, string description,
+            [FromQuery] HashSet<int> characters, [FromQuery] HashSet<int> movies)
         {
             await using var db = new PersonContext();
             var person = new Person
@@ -71,8 +71,8 @@ namespace MovieWebsite.Persons.Controllers
         /// <returns>Returns id of added Person if success, set null if you dont want to update</returns>
         [Authorize]
         [HttpPost("UpdatePerson")]
-        public async Task<IActionResult> UpdatePerson(long id, string? name, string? originalName, long? posterId, string? description,
-            [FromQuery] HashSet<long>? characters, [FromQuery] HashSet<long>? movies)
+        public async Task<IActionResult> UpdatePerson(int id, string? name, string? originalName, int? posterId, string? description,
+            [FromQuery] HashSet<int>? characters, [FromQuery] HashSet<int>? movies)
         {
             //this function should not be run very often, so efficiency is not so important
             await using var db = new PersonContext();
